+++ js/settings/personal.js
@@ -32,57 +32,6 @@ $(document).ready(function () {
         });
     });
 
-    $('#sonos').on('click', function () {
-        var user_value;
-        if ($('#sonos').prop('checked')) {
-            user_value = 'checked';
-        }
-        else {
-            user_value = '';
-        }
-        $.ajax({
-            type: 'GET',
-            url: OC.generateUrl('apps/audioplayer/setvalue'),
-            data: {
-                'type': 'sonos',
-                'value': user_value
-            },
-            success: function () {
-                OC.Notification.showTemporary(t('audioplayer', 'saved'));
-            }
-        });
-    });
-
-    $('#sonos_controller_submit').on('click', function () {
-        var user_value = $('#sonos_controller').val();
-        $.ajax({
-            type: 'GET',
-            url: OC.generateUrl('apps/audioplayer/setvalue'),
-            data: {
-                'type': 'sonos_controller',
-                'value': user_value
-            },
-            success: function () {
-                OC.Notification.showTemporary(t('audioplayer', 'saved'));
-            }
-        });
-    });
-
-    $('#sonos_smb_path_submit').on('click', function () {
-        var user_value = $('#sonos_smb_path').val();
-        $.ajax({
-            type: 'GET',
-            url: OC.generateUrl('apps/audioplayer/setvalue'),
-            data: {
-                'type': 'sonos_smb_path',
-                'value': user_value
-            },
-            success: function () {
-                OC.Notification.showTemporary(t('audioplayer', 'saved'));
-            }
-        });
-    });
-
     /*
  * Collection path
  */
@@ -130,21 +79,5 @@ $(document).ready(function () {
             $('#browser_no').html(nsupported_types);
         }
     });
-
-    $.ajax({
-        type: 'POST',
-        url: OC.generateUrl('apps/audioplayer/sonosdevices'),
-        data: {},
-        success: function (jsondata) {
-            $(jsondata).each(function (i, el) {
-                $('#sonos_controller').append($('<option>', {
-                    value: el[0],
-                    text: el[2][0]
-                }));
-            });
-            $('#sonos_controller').val($('#sonos_current').val());
-        }
-    });
-
 
 });